# perf: specialize `PartialOrd` for `UnsignedInteger`