@Oppen 2023-06-15
> Nice! What about the PartialEq and Clone derived implementations? Are those efficient?

Haven't checked, but those didn't come up in flamegraphs as obviously as this.